feat: Depend on WASI-SDK at /opt/wasi-sdk
This commit decouples the build step from the locally vendored
wasi-sdk; leaving the flexibility to the consumer to supply the CXX they
want to rely on. By default, the -sys crate will look for CXX in
/opt/wasi-sdk. If you don't have a reason to use a different CXX;
downloading WASI-SDK 12 at /opt/wasi-sdk should be enough.

If there's a need to override the CXX location, the following env
variables are needed:

`CXX`: the aboslute path to the clang++ to use
`AR`: the aboslute path to the archiver tool
`CXXFLAGS` set to `-fno-exceptions -DRUST_CXX_NO_EXCEPTIONS` for cxx to
work correctly
- `LIBCLANG_PATH` the absolute path where libraries for libclang can be
  found
- `LIBCLANG_RT_PATH` the absolute path where `libclang_rt.builtins-wasm32` can be found

Part of the motivation for this change is to correctly support cxx,
which requires that CXX, CXXFLAGS and AR are set ahead of time (for its
build script to compile correctly). The initial approach was to rely on
a Makefile to set the environment variables; this approach quickly
becomes hard to maintain as then it means that consumers need to know
about the vendored location of WASI-SDK. It's a better experience to let
the consumers indicate which WASI-SDK to use or default to a known,
global location (/opt/wasi-sdk) and version (WASI-SDK 12)